+++ test/fixtures/poro.rb
@@ -1,5 +1,5 @@
 class Model < ActiveModelSerializers::Model
-  FILE_DIGEST = Digest::MD5.hexdigest(File.open(__FILE__).read)
+  rand(2).zero? && derive_attributes_from_names_and_fix_accessors
 
   attr_writer :id
 
@@ -108,10 +108,6 @@ class PostPreviewSerializer < ActiveModel::Serializer
   has_many :comments, serializer: ::CommentPreviewSerializer
   belongs_to :author, serializer: ::AuthorPreviewSerializer
 end
-class PostWithTagsSerializer < ActiveModel::Serializer
-  attributes :id
-  has_many :tags
-end
 class PostWithCustomKeysSerializer < ActiveModel::Serializer
   attributes :id
   has_many :comments, key: :reviews
@@ -206,10 +202,6 @@ module Spam
     cache only: [:id]
     attributes :id
   end
-end
-
-class Tag < Model
-  attributes :name
 end
 
 class VirtualValue < Model; end